+++ quiltcore/namespace.py
@@ -8,10 +8,12 @@ class Namespace(ResourcePath):
     Namespace of Manifests by Hash
     list/get returns a specific Manifest
     """
+    SEP = "/"
 
     def __init__(self, path: Path, **kwargs):
         super().__init__(path, **kwargs)
         self.manifests = kwargs[self.KEY_MAN]
+        self.name = self.pkg_name()
 
     def hash(self, tag: str = ResourcePath.TAG_DEFAULT) -> str:
         hash_file = self.path / tag
@@ -19,5 +21,9 @@ class Namespace(ResourcePath):
 
     def _child_path(self, key: str) -> Path:
         """Return the path for a child resource."""
+        # TODO: match on partial hashes
         hash = self.hash(key)
         return self.manifests / hash
+
+    def pkg_name(self) -> str:
+        return self.SEP.join(self.path.parts[-2:])