+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # CHANGELOG.md
 
-## 0.3.0 (UNRELEASED)
+## 0.3.1 (UNRELEASED)
+
+QuiltPlus fixes
+
+- Namespace.pkg_name() returns the package name
+- Load Manifests lazily
+- Allow getting not-yet-existing Namespaces from a Registry
+
+## 0.3.0 (2023-07-27)
 
 - quiltspec
 - push new packages